uucore: Move extendedbigdecimal to its own feature
This will be needed later on so that we can split format and
parse features.
